Repeat item shuffle until ROM is completable.
Added code to check the logic that would applied to the ROM. The code
just repeatedly shuffles until the logic works or it reaches a maximum
number of attempts.

An error message is displayed if the logic checker has to give up. This
should only really display if the user logic is impossible to complete.