Update Asm patch files, write D-Pad configuration as a struct
This has a side effect: if using an older version to apply a patch from this
version or newer, the D-Pad configuration won't be affected and will instead be
whatever the creator of the patch configured it as. This is because the older
software will overwrite the magic number of the D-Pad configuration struct
instead of the buttons fields.